+++ contracts/cw-cyber-airdrop/schema/execute_msg.json
@@ -95,17 +95,20 @@
         "claim": {
           "type": "object",
           "required": [
-            "claim_amount",
-            "claim_msg",
-            "proof",
-            "signature"
+            "gift_amount",
+            "gift_claiming_address",
+            "nickname",
+            "proof"
           ],
           "properties": {
-            "claim_amount": {
+            "gift_amount": {
               "$ref": "#/definitions/Uint128"
             },
-            "claim_msg": {
-              "$ref": "#/definitions/ClaimMsg"
+            "gift_claiming_address": {
+              "type": "string"
+            },
+            "nickname": {
+              "type": "string"
             },
             "proof": {
               "description": "Proof is hex-encoded merkle proof.",
@@ -113,9 +116,6 @@
               "items": {
                 "type": "string"
               }
-            },
-            "signature": {
-              "$ref": "#/definitions/Binary"
             }
           }
         }
@@ -129,59 +129,21 @@
       ],
       "properties": {
         "release": {
-          "type": "object"
+          "type": "object",
+          "required": [
+            "gift_address"
+          ],
+          "properties": {
+            "gift_address": {
+              "type": "string"
+            }
+          }
         }
       },
       "additionalProperties": false
     }
   ],
   "definitions": {
-    "Binary": {
-      "description": "Binary is a wrapper around Vec<u8> to add base64 de/serialization with serde. It also adds some helper methods to help encode inline.\n\nThis is only needed as serde-json-{core,wasm} has a horrible encoding for Vec<u8>",
-      "type": "string"
-    },
-    "ClaimMsg": {
-      "type": "object",
-      "required": [
-        "avatar_cid",
-        "gift_claiming_address",
-        "gift_claiming_address_type",
-        "nickname",
-        "relay_reward",
-        "target_address"
-      ],
-      "properties": {
-        "avatar_cid": {
-          "type": "string"
-        },
-        "gift_claiming_address": {
-          "type": "string"
-        },
-        "gift_claiming_address_type": {
-          "$ref": "#/definitions/ClaimerType"
-        },
-        "nickname": {
-          "type": "string"
-        },
-        "relay_reward": {
-          "$ref": "#/definitions/Decimal"
-        },
-        "target_address": {
-          "type": "string"
-        }
-      }
-    },
-    "ClaimerType": {
-      "type": "string",
-      "enum": [
-        "ethereum",
-        "cosmos"
-      ]
-    },
-    "Decimal": {
-      "description": "A fixed-point decimal value with 18 fractional digits, i.e. Decimal(1_000_000_000_000_000_000) == 1.0\n\nThe greatest possible value that can be represented is 340282366920938463463.374607431768211455 (which is (2^128 - 1) / 10^18)",
-      "type": "string"
-    },
     "Uint128": {
       "description": "A thin wrapper around u128 that is using strings for JSON encoding/decoding, such that the full u128 range can be used for clients that convert JSON numbers to floats, like JavaScript and jq.\n\n# Examples\n\nUse `from` to create instances of this and `u128` to get the value out:\n\n``` # use cosmwasm_std::Uint128; let a = Uint128::from(123u128); assert_eq!(a.u128(), 123);\n\nlet b = Uint128::from(42u64); assert_eq!(b.u128(), 42);\n\nlet c = Uint128::from(70u32); assert_eq!(c.u128(), 70); ```",
       "type": "string"
