+++ contracts/cw-cyber-airdrop/src/msg.rs
@@ -1,7 +1,7 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use cosmwasm_std::{Binary, Decimal, Uint128, Uint64};
+use cosmwasm_std::{Decimal, Uint128, Uint64};
 use cw_utils::Expiration;
 
 #[derive(Serialize, Deserialize, JsonSchema)]
@@ -38,48 +38,13 @@ pub enum ExecuteMsg {
     },
     /// Claim does not check if contract has enough funds, owner must ensure it.
     Claim {
-        claim_msg: ClaimMsg,
-        signature: Binary,
-        claim_amount: Uint128,
+        nickname: String,
+        gift_claiming_address: String,
+        gift_amount: Uint128,
         /// Proof is hex-encoded merkle proof.
         proof: Vec<String>,
     },
-    Release {},
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct ClaimMsg {
-    pub nickname: String,
-    pub avatar_cid: String,
-    pub gift_claiming_address_type: ClaimerType,
-    pub gift_claiming_address: String,
-    pub target_address: String,
-    pub relay_reward: Decimal,
-}
-
-/*
-impl ToString for ClaimMsg {
-    fn to_string(&self) -> String {
-        format!("\{{}\}", self.target_address, )
-    }
-}
-
- */
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum ClaimerType {
-    Ethereum,
-    Cosmos,
-}
-
-impl ToString for ClaimerType {
-    fn to_string(&self) -> String {
-        match self {
-            ClaimerType::Ethereum => String::from("ethereum"),
-            ClaimerType::Cosmos => String::from("cosmos"),
-        }
-    }
+    Release { gift_address: String},
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -120,9 +85,32 @@ pub struct IsClaimedResponse {
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct ReleaseStateResponse {
+    pub address: String,
     pub balance_claim: Uint128,
     pub stage: Uint64,
     pub stage_expiration: Expiration,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct PassportSignedQuery {
+    pub nickname: String,
+    pub address: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct PassportSignedResponse {
+    pub signed: bool,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct PassportAddressQuery {
+    pub nickname: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, PartialEq, JsonSchema, Debug)]
+#[serde(rename_all = "snake_case")]
+pub struct PassportAddressResponse {
+    pub address: String,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
