+++ contracts/cw-cyber-passport/src/query.rs
@@ -1,6 +1,6 @@
 use cosmwasm_std::{Deps, StdResult};
 use cw721_base::state::TokenInfo;
-use crate::msg::{ConfigResponse, PortidResponse, AddressResponse};
+use crate::msg::{ConfigResponse, PortidResponse, AddressResponse, SignatureResponse};
 use crate::state::{CONFIG, NICKNAMES, PassportContract, PassportMetadata, PORTID};
 
 pub fn query_config(deps: Deps) -> StdResult<ConfigResponse> {
@@ -27,11 +27,28 @@ pub fn query_passport_by_nickname(deps: Deps, nickname: String) -> StdResult<Tok
 }
 
 pub fn query_address_by_nickname(deps: Deps, nickname: String) -> StdResult<AddressResponse> {
-    let address_portid = NICKNAMES.may_load(deps.storage, &nickname)?;
-    Ok(AddressResponse { address: address_portid.unwrap().address.into() })
+    let address_portid = NICKNAMES.load(deps.storage, &nickname)?;
+    Ok(AddressResponse { address: address_portid.address.into() })
 }
 
 pub fn query_portid(deps: Deps) -> StdResult<PortidResponse> {
     let portid = PORTID.load(deps.storage)?;
     Ok(PortidResponse { portid: portid.into() })
+}
+
+pub fn query_nickname_address_signed(
+    deps: Deps,
+    nickname: String,
+    address: String
+) -> StdResult<SignatureResponse> {
+    let address_portid = NICKNAMES.load(deps.storage, &nickname)?;
+    let cw721_contract = PassportContract::default();
+    let token_info = cw721_contract
+        .tokens
+        .load(deps.storage, address_portid.portid.as_str())?;
+    let mut result = false;
+    if token_info.clone().extension.addresses.is_some() {
+        result = token_info.clone().extension.addresses.unwrap().iter().any(|i| i.as_ref() == address);
+    }
+    Ok(SignatureResponse { signed:  result })
 }
